bin: Add kbsecret-pass.
kbsecret-pass is a lot like kbsecret-login, except that it only
retrieves the password field (instead of the whole login record)
and allows you to shunt the password directly into the system
clipboard via the -c/--clipboard flag. It also only retrieves
one record at a time.

gemspec: Add clipboard as a dependency.