bin: Add kbsecret-raw-edit.
kbsecret-raw-edit allows the user to edit the raw JSON of the requested
record. This is a bit dangerous (the user could break the JSON structure),
but the use case here is really just me experimenting and possibly users
who are more comfortable changing text than they are using the current
command-line tools.