config: Add Config.deconfigure_session for "hiding" a session.
This just removes a particular session's configuration from the global
configuration, making it functionally invisible to the rest of kbsecret.
Sessions that are deconfigured via deconfigure_session can be re-configured
without loss of records via Config.configure_session (or kbsecret new-session).

session: Add Session#unlink! for completely deleting a session.

This is the destructive complement to Config.deconfigure_session.
When called, this removes the entire session directory and all records
within. It does *not* remove the session's configuration entry, meaning
that future attempts to access the session will create a blank root
of the same name.