kbsecret-new: Replace -i/--interactive with -a/--args.
The previous commit makes it hard to write scripts that use the argument form
of `kbsecret new` in legitimate ways. To avoid this, the -a/--args options
now allow the default (interactive) mode to be overridden in favor of
supplying fields as arguments in any situation.

Additionally, an interactive terminal is not available, the arguments form
is still defaulted to. It is only when an interactive terminal is available
*and* -a/--args has not been passed that interactive input occurs.