lib: Be more consistent about record label type.
Public methods should always take either a string or a symbol as an
identifier for a given record, even though record labels are represented
as strings internally. This is because it's a lot nicer to ask
for the :github or other well-recognized label than for "github",
but at the same time :"my-complex-label" is ugly.

Session label types are still inconsistent for the time being, so
always use symbols to be safe.