cli: Convert KBSecret::CLI into a class.
KBSecret::CLI can now be instantiated, with a block provided to
introduce the Slop (and Dreck) state previously created by the
KBSecret::CLI.slop method.

This was done primarily to make trailing argument handling via
Dreck much nicer, as the alternative would have been a bunch
of this littered in each command:

	Dreck.parse opts.args do
		# ...
	end

With this in place, KBSecret::CLI.slop is now considered deprecated
(and will be removed before the first stable release).