record: Support for field sensitivity.
By default, all fields in a record created via `data_field` are marked
as "sensitive," meaning that userspace programs should consider their
contents private or otherwise not safe to echo to the user by default.

The `sensitive?(field)` method is provided to test whether a particular
field is sensitive or not. It's available both as a class method
(e.g., `Login.sensitive?(:password)`) and as an instance method
via delegation. The `data_fields` method has been similarly delegated,
for convenience.