lib: Preliminary support for secret generation
This commit introduces KBSecret::Generator, which will serve
as the basis for all secret generation in KBSecret.

The Generator class is straightforward: It gets instantiated with the name
of a generator "profile", which is defined under the "generators" key
in the user's config. By default, the "default" profile is used,
which defines a key length of 16 bytes and a key format of "hex"
(i.e., a hexadecimal string).

Once instantiated, the Generator#secret method is called at-will to generate
a new secret matching the profile's parameters.

In addition, this commit introduces a few new exception classes, some
documentation fixes, and so on.