lib: Remove custom record initializers
This removes custom initialize methods in each record class
in favor of testing the number of external fields expected
and generating the body of each record directly in
Session#add_record.

The populate_internal_fields method is introduced as a replacement
for custom initializers, and can be used by record classes to fill
in default values (or do anything really) for their internal fields.