lib: Add support for per-command configuration
This commit introduces per-command configuration to `kbsecret` via
the "commands.ini" file. "commands.ini" is an INI-formatted file
with the following layout:

    [command]
    args = --option1 -abc --option2 value

Where "command" is the short name of a `kbsecret` command (e.g.,
`kbsecret-command`), and "args" is a string of arguments (following
shell quoting rules) that get supplied to `kbsecret-command` by default.

Keys other than "args" under a command-section are not currently
used for anything, but can be accessed via `Config.command(cmd)`.

Sectionless keys (i.e., those at the very beginning) are currently
unused. This might change in the future.

Closes #2.