kbsecret-new: Deprecate -a/--args for -x/--terse
This commit marks the -a/--args flag to `kbsecret new` as deprecated,
and spits a warning at the user should they try to use it. The
functionality of -a/--args will be removed entirely by 1.0.

The new -x/--terse option serves as a replacement for -a/--args,
accepting a tersely-formatted string of fields on standard input.
The behavior of this flag can be modified via -i/--ifs.
Compare this to `kbsecret login` and `kbsecret dump-fields`.

In short, usages like this:

    $ kbsecret new login -a foo bar baz

can be turned into this:

    $ kbsecret new login -x foo <<< "bar:baz"

Which isn't substantially longer, and is much safer.