session: Keybase team support
This commit adds support for Keybase teams to KBSecret's concept
of a session.

Team-based sessions have a `:team` key in their configuration
instead of a normal `:users` key. This key is assumed to be a
valid Keybase team name, accessible at `/keybase/team/:team`
(or wherever your mountpoint is).

No validation of the team or the user's permissions within it is
currently done.