more KBFS health checks
Hey there! Thanks for the awesome project!

We've seen several bug reports where `kbsecret` users had process written into a dir that's supposed to be the KBFS mount point, while it's not mounted. It seems it's because `kbsecret` only checks for existence of `/keybase`, but doesn't make sure it's actually mounted. To mitigate that, this PR checks for `/keybase/.kbfs_status` existence, which is a (hidden) special file that's very unlikely to get written into. In addition (suggested by @strib), check content of `/keybase/.kbfs_status` to make sure user is logged in.

I know close to nothing about ruby, so please bear with me for any stupid mistakes, and feel free to throw it away if there's a more idiomatic/cleaner way to do this!