test: Run Keybase on the CI
These changes eliminate the discrepancies between testing locally
(with a full Keybase installation) and testing on the CI
(with stubs for Keybase and a fake KBFS mount).

This reduces the amount of hacky stub code needed to run the tests,
*and* allows for the running of CLI tests on the CI.

Implementation notes:

* The `setup.expect` script needs an environment variable,
`KBSECRETCI_PAPERKEY`, to log into Keybase. This has been set in the
Travis CI webUI and made secret, to prevent accidental disclosure.

* Travis CI now runs only one test process at a time. This is because of
how device provisioning works on Keybase - if a device is already
provisioned at `login` time, `keybase login` will ask for interactive
confirmation from that device. Getting that to work on the CI would
involve a lot of hackiness, so instead we wait for each device to
deprovision itself before starting the next test process.