lib: Invariant cleanup
This commit moves some necessary invariant checks out of the CLI and
into the library itself, where they belong. In particular,
`Session#add_record` and `Session#import_record` now do their own
overwrite checks, and throw an exception to indicate when a record
addition or import would cause an overwrite.

These exceptions can be silenced via the `overwrite` keyword argument
for both methods, which is filled in the CLI by the state of the
`--force` flag.

This commit also adds unit tests to cover this new behavior.